requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# LLM and AI
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.13
langgraph==0.0.20

# Vector Search and Embeddings
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

